moved tests for new io.misc
